Various audio fixes
  1. GP4/GP5 tremolo bar depth (src/parser/gp345/song_parser.rs:501) — value / GP_BEND_SEMITONE * 2.0 became value / (GP_BEND_SEMITONE *
  2.0), matching TuxGuitar's value / (25f * 2f). Dives are now 4× shallower (correct): in the regenerated golds the old output was slamming
  dives into the pitch-wheel floor of 0 (−12 semitones), while the new minimum is a plausible 20.

  2. Slide interpolation (src/audio/midi_builder/builder.rs) — the tone computation now does the final division in f32 and uses TuxGuitar's
  exact truncate-then-add semantics (DEFAULT_BEND as i32 + (tone * 5.5) as i32), which also matches Java's toward-zero truncation on
  downward slides. I hand-verified against the reference formula using the Guthrie Govan gold: the descending slide at tick 34200 now emits
  61, 57, 53 (TuxGuitar-exact) instead of the old staircase 64, 58, 53, and the ascending one ramps evenly 66, 69, 72, 75 instead of 64, 69,
  69, 75.

  3. Empty-points guards (song_parser.rs) — the two bend call sites (GP3 and GP4/5) and the GP4/5 tremolo-bar site now only set the effect
  when the parsed point list is non-empty, mirroring TuxGuitar's isEmpty() checks. This also closes the interaction with the priority chain
  where a zero-point bend would have suppressed vibrato.

+++ src/audio/midi_builder/builder.rs
@@ -416,10 +416,11 @@ impl MidiBuilder {
                     let length: i32 = (tick2 - tick1) as i32;
                     let points = length / (QUARTER_TIME / 8) as i32;
                     for p_offset in 1..=points {
-                        let tone = ((length / points) * p_offset) * distance / length;
-                        let bend = DEFAULT_BEND + (tone as f32 * DEFAULT_BEND_SEMI_TONE * 2.0);
+                        let tone = ((length / points) * p_offset * distance) as f32 / length as f32;
+                        let bend =
+                            DEFAULT_BEND as i32 + (tone * DEFAULT_BEND_SEMI_TONE * 2.0) as i32;
                         let bend_tick = tick1 as i32 + (length / points) * p_offset;
-                        self.add_pitch_bend(bend_tick as u32, track_id, channel_id, bend as i32);
+                        self.add_pitch_bend(bend_tick as u32, track_id, channel_id, bend);
                     }
 
                     // normalise the bend
